To-Do list works.
Deleted the template HTML and updated JS to now accept new items to the
list, which then renders them out in the unordered list below the
input. I used another classmate’s code for reference, because I got
confused on how to push to the array via the form. It all works now!