added a data variable to item list
This is an array that has to do with the category list. I renamed
category list to item list. I don’t know how to handle the author label
from the tutorial. Should it be the selected category and the priority,
them the msg item.  Or, should I even bother with it?